Enhance demo mode functionality and update documentation
- Added new entries to .gitignore for setup and Docker fixes.
- Updated README with revised demo mode instructions and access points.
- Modified config to set CSV seed path and demo mode flag.
- Implemented demo mode checks in the demo router to ensure proper functionality.
- Improved error handling and logging for CSV file loading in demo mode.
- Updated frontend to include a dedicated Demo page and navigation link.

+++ app/services/recommender.py
@@ -127,7 +127,8 @@ class Recommender:
                 print(f"  ✅ Feature columns: {feature_columns}")
             
             # Build FAISS index from pre-trained embeddings
-            X = np.ascontiguousarray(embeddings)
+            # IMPORTANT: FAISS requires float32 dtype
+            X = np.ascontiguousarray(embeddings.astype('float32'))
             self.feature_vectors = X.copy()  # Store original for MMR
             
             # Normalize for cosine similarity
